[models] delete base.py and added model
