--- django_freeradius/base/base.py
@@ -1,16 +1,0 @@
-from django.db import models
-from django.utils.translation import ugettext_lazy as _
-
-from model_utils.fields import AutoCreatedField, AutoLastModifiedField
-
-
-class TimeStampedEditableModel(models.Model):
-    """
-    An abstract base class model that provides self-updating
-    ``created`` and ``modified`` fields.
-    """
-    created = AutoCreatedField(_('created'), editable=True)
-    modified = AutoLastModifiedField(_('modified'), editable=True)
-
-    class Meta:
-        abstract = True
